chore: remove unnecessary custom hooks, refactor home and payments pages
