+++ src/Hooks/index.ts
@@ -1,13 +1,10 @@
 export { default as useLanguage } from "./useLanguage/useLanguage";
 export { default as useMainButton } from "./useMainButton/useMainButton";
 export { default as useRandomCards } from "./useRandomCards/useRandomCards";
-export { default as useLowBalancePopup } from "./useLowBalancePopup/useLowBalancePopup";
 export { default as useMainButtonHandler } from "./useMainButtonHandler/useMainButtonHandler";
 export { default as useSettingsButton } from "./useSettingsButton/useSettingsButton";
 export { default as useReadings } from "./useReadings/useReadings";
-export { default as useNavigateToQuestion } from "./useNavigateToQuestion/useNavigateToQuestion";
 export { default as useBackButton } from "./useBackButton/useBackButton";
 export { default as useSendSpread } from "./useSendSpread/useSendSpread";
-export { default as useCoinsPurchase } from "./useCoinsPurchase/useCoinsPurchase";
 export { default as useDisableAdsPurchase } from "./useDisableAdsPurchase/useDisableAdsPurchase";
 export { default as useAds } from "./useAds/useAds";
